Implement label boundary rendering (de)serialization (#686)

+++ src/main/java/org/embl/mobie/viewer/command/LabelRenderingConfiguratorCommand.java
@@ -23,7 +23,7 @@ public class LabelRenderingConfiguratorCommand implements BdvPlaygroundActionCom
 	boolean showAsBoundary;
 
 	@Parameter ( label = "Boundary thickness", style="format:#.00" )
-	double boundaryThickness = 1.0;
+	float boundaryThickness = 1.0F;
 
 	@Override
 	public void run()
